requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyJWT==2.6.0
pyTelegramBotAPI==4.11.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
